pyproject.toml
[tool.poetry]
name = "lovejoy"

version = "1.0.0"

description = " "

authors = [
    "Ali Ahmed alimohdapg@gmail.com",
]

[tool.poetry.dependencies]
python = ">=3.8,<4"
django = "^4.1.4"
dj-database-url = "^1.0.0"
psycopg2-binary = "^2.9.5"
django-environ = "^0.9.0"
django-crispy-forms = "^1.14.0"
whitenoise = "^6.2.0"
gunicorn = "^20.1.0"
pillow = "^9.3.0"
setuptools = "^65.6.3"
dropbox = "^11.36.0"
django-storages = "^1.13.1"
django-recaptcha = "^3.0.0"
certifi = "^2022.12.7"